+++ CHANGES.rst
@@ -5,5 +5,6 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
-- Initial release.
+- Initial release. Support of Video service for the new wildcard.media's
+  WildcardVideo content type.
   [arsenico13]